# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * YMKCOE College Website API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Course {
  id: number;
  name: string;
  department: string;
  /** UG or PG */
  type: string;
  duration: string;
  seats: number;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  eligibility?: string | null;
  /** @nullable */
  fees?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  isActive?: boolean;
  createdAt: string;
}

export interface CourseInput {
  name: string;
  department: string;
  type: string;
  duration: string;
  seats: number;
  description?: string;
  eligibility?: string;
  fees?: string;
  imageUrl?: string;
  isActive?: boolean;
}

export interface CourseUpdate {
  name?: string;
  department?: string;
  type?: string;
  duration?: string;
  seats?: number;
  description?: string;
  eligibility?: string;
  fees?: string;
  imageUrl?: string;
  isActive?: boolean;
}

export interface FacultyMember {
  id: number;
  name: string;
  department: string;
  designation: string;
  /** @nullable */
  qualification?: string | null;
  /** @nullable */
  specialization?: string | null;
  /** @nullable */
  experience?: number | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  isHOD?: boolean;
  createdAt: string;
}

export interface FacultyInput {
  name: string;
  department: string;
  designation: string;
  qualification?: string;
  specialization?: string;
  experience?: number;
  email?: string;
  phone?: string;
  photoUrl?: string;
  bio?: string;
  isHOD?: boolean;
}

export interface FacultyUpdate {
  name?: string;
  department?: string;
  designation?: string;
  qualification?: string;
  specialization?: string;
  experience?: number;
  email?: string;
  phone?: string;
  photoUrl?: string;
  bio?: string;
  isHOD?: boolean;
}

export interface AdmissionLead {
  id: number;
  name: string;
  email: string;
  phone: string;
  courseInterest: string;
  /** @nullable */
  qualification?: string | null;
  /** @nullable */
  message?: string | null;
  /** new | contacted | enrolled | rejected */
  status: string;
  createdAt: string;
}

export interface AdmissionLeadInput {
  name: string;
  email: string;
  phone: string;
  courseInterest: string;
  qualification?: string;
  message?: string;
}

export interface AdmissionLeadUpdate {
  status?: string;
  message?: string;
}

export interface NewsItem {
  id: number;
  title: string;
  content: string;
  /** announcement | event | achievement | news */
  category: string;
  /** @nullable */
  imageUrl?: string | null;
  publishedAt: string;
  isPinned?: boolean;
  createdAt: string;
}

export interface NewsInput {
  title: string;
  content: string;
  category: string;
  imageUrl?: string;
  publishedAt?: string;
  isPinned?: boolean;
}

export interface NewsUpdate {
  title?: string;
  content?: string;
  category?: string;
  imageUrl?: string;
  publishedAt?: string;
  isPinned?: boolean;
}

export interface MediaItem {
  id: number;
  title: string;
  /** image | video */
  type: string;
  url: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** campus | events | labs | sports | cultural */
  category: string;
  createdAt: string;
}

export interface MediaInput {
  title: string;
  type: string;
  url: string;
  thumbnailUrl?: string;
  category: string;
}

export interface Placement {
  id: number;
  studentName: string;
  company: string;
  /** e.g. 5.5 LPA */
  package: string;
  /** @nullable */
  role?: string | null;
  department: string;
  year: number;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  testimonial?: string | null;
  createdAt: string;
}

export interface PlacementInput {
  studentName: string;
  company: string;
  package: string;
  role?: string;
  department: string;
  year: number;
  logoUrl?: string;
  testimonial?: string;
}

export interface PlacementUpdate {
  studentName?: string;
  company?: string;
  package?: string;
  role?: string;
  department?: string;
  year?: number;
  logoUrl?: string;
  testimonial?: string;
}

export interface ContactSubmission {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  subject: string;
  message: string;
  createdAt: string;
}

export interface ContactInput {
  name: string;
  email: string;
  phone?: string;
  subject: string;
  message: string;
}

export interface AdminCredentials {
  username: string;
  password: string;
}

export interface AdminSession {
  id: number;
  username: string;
  isAdmin: boolean;
}

export interface DashboardStats {
  totalCourses: number;
  totalFaculty: number;
  totalStudents: number;
  totalPlacements: number;
  totalLeads: number;
  recentNews: number;
  pendingLeads: number;
}

export type PlacementStatsTopCompaniesItem = {
  company: string;
  count: number;
};

export type PlacementStatsByYearItem = {
  year: number;
  count: number;
};

export type PlacementStatsByDepartmentItem = {
  department: string;
  count: number;
};

export interface PlacementStats {
  totalPlaced: number;
  averagePackage: string;
  highestPackage: string;
  topCompanies: PlacementStatsTopCompaniesItem[];
  byYear: PlacementStatsByYearItem[];
  byDepartment: PlacementStatsByDepartmentItem[];
}

export type GetCoursesParams = {
department?: string;
type?: string;
};

export type GetFacultyParams = {
department?: string;
};

export type GetAdmissionLeadsParams = {
status?: string;
};

export type GetNewsParams = {
category?: string;
};

export type GetMediaItemsParams = {
category?: string;
};

export type GetPlacementsParams = {
year?: number;
department?: string;
};