# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: YMKCOE College Website API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCredentials"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Invalid credentials
...
  /admin/me:
    get:
      operationId: getAdminMe
      tags: [admin]
      summary: Get current admin session
      responses:
        "200":
          description: Current session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Not authenticated

  # ─── STATS (wow endpoints) ────────────────────────────────────────────────────
...
components:
  schemas:
...
    AdminCredentials:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }
    AdminSession:
      type: object
      required: [id, username, isAdmin]
      properties:
        id: { type: integer }
        username: { type: string }
        isAdmin: { type: boolean }

    # ─── STATS ─────────────────────────────────────────────────────────────────
...